+++ changes.md
@@ -4,6 +4,7 @@
 1. Fix problem with curved TGraph drawings (#218)
 2. Fix problem with logy/logz drawing in lego plots
 3. Provide x3dscNNN and y3dscNNN draw option for histogram to resize x/y axis
+4. Provide JSROOT.Painter.setDefaultDrawOpt to change default draw option
 
 
 ## Changes in 6.2.0